Add Most played player in game overview, Add currency in travel expense, Update Hr Dashboard

+++ Backend/src/main/java/com/roima/HRMS/entities/TravelDetail.java
@@ -32,7 +32,7 @@ public class TravelDetail {
 
     @PositiveOrZero
     @Column(name = "max_amout_per_day",nullable = false)
-    private Long maxAmoutPerDay;
+    private Double maxAmoutPerDay;
 
     @Column(name = "start_date",nullable = false)
     private LocalDateTime startDate;
